pwhm: wpa_supplicant: Limit 6 GHz scan to preferred scanning channels
On repeaters with a 6 GHz uplink, the time to scan the entire 6 GHz band to find
the gateway can be long. Here we introduce a configurable parameter to limit the
scan to preferred channels, as defined in 80211ax-2021 section 26.17.2.3. The
parameter is false by default as safer option. In deployments where the backhaul
links are known to use PSCs, the option should be enabled to allow faster
establishment of the 6 GHz uplink.

GitOrigin-RevId: d2f0b69295c7aed792a4812c421f77028a55b513

Implements: PPM-3681
References: FEAT-117
Signed-off-by: Quentin Feraboli <quentin.feraboli@softathome.com>
Signed-off-by: Ahmed Amine HAOUES <ahmed-amine.haoues_ext@softathome.com>