pwhm: ep: fix: Missing target chanspec update after EP connection
Problem:
If Station connected event is missed, a chanSpec configuration
mismatch is not detected.

Fix:
If the endpoint is already connected when the start connection event is received,
perform a chanspec refresh. This prevents mismatches between BH and FH if the station connected event is missed.

GitOrigin-RevId: c75e560d52e965fe9783e8043f22810046960828

Implements: PPM-3681
References: FEAT-117
Signed-off-by: Quentin Feraboli <quentin.feraboli@softathome.com>
Signed-off-by: Ahmed Amine HAOUES <ahmed-amine.haoues_ext@softathome.com>