wld: Calculate the Interference Factor from the channel survey report after each scan is completed.
Problem:
The multiApPreference value is currently always set to the default of 14 for all operable channels,
which does not reflect actual channel conditions.

Fix:
Obtain the channel survey report from the driver and compute the Interference Factor for each channel
immediately after the scan is completed.This values should then be stored in the wld_surveyReport_t structure.
The Interference Factor will later be used in the channel ranking process in prplMesh,
which in turn determines the multiApPreference value.

Interference Factor Calculation:
The interference factor is the ratio of observed busy time to the total time spent on the channel.
This ratio is then amplified by the relative noise floor, the observed noise floor on the current
channel compared to the lowest noise floor observed across the entire band.

Closes: PPW-531

Signed-off-by: pavani sireesha <kareti-pavani.sireesha-reddy@capgemini.com>