pwhm: fsm: fix: Extender is not pairing with router in WPS
Problem: After flashing a new image, and performing a factory reset on both router and extender. extender is not pairing with router in WPS.

Cause: When s_syncOnEpConnected is called, a mismatch is detected between currently used bw of the connected ep and the ep target bssid scanned one. This is because
a specific check previously introduced in the target scanned bw calculation, forcing it to 20 BW for the 2,4 GHz band. This causes a loop and a permanent reconfigure of
the ep.

Fix: add a post condition in the s_syncOnEpEconnected function after the target scanned bw calculation, to avoid applying the target scanned bw in case lower than the current bw in use.

GitOrigin-RevId: 2bed12853469035cd756ff7a40e15b957ffc2126

Implements: PPM-3681
References: FEAT-117
Signed-off-by: Quentin Feraboli <quentin.feraboli@softathome.com>
Signed-off-by: Ahmed Amine HAOUES <ahmed-amine.haoues_ext@softathome.com>