mfn_wrad_sync: synchronize security parameters with secDmn's running context
Issue:
The security of the VAP is being reset to the default choice during every
radio sync.

Cause:
When a radio sync is triggered, the security parameter (`wpa_key_mgmt`)
for all the VAPs in the respective Radio is set to the Primary VAP’s value.
This affects the security of all the VAPs. The respective `hostap.conf` file
is updated correctly unlike the running context.

Solution:
Set the AP Sec sync FSM (`s_doSetApSec`) for a Radio Sync action.

Fixes: PPW-889

Signed-off-by: Sagar Reddy <snandanuri@maxlinear.com>