pwhm: security config: add WPA3 Compatibility Mode
WPA3-Personal-Compatibility abbrev. to WPA3-CM

WPA3-CM specifies a precise configuration for
the contents of the RSN Override 1/2
and RSNX Override IEs

As such, WPA3-CM behaves like other WPAx security modes.

Add the corresponding code in the function
that generates the configuration map;

Since the support for the RSN Override mechanism
is not yet available as a driver-reported capability,
add a flag to 'manually' enable the feature.
The flag adds the WPA3-CM mode to Security.ModesSupported
list.

Add WPA3-CM to validation cycle of 11BE features:
wld_ap_sec_checkSharedSecConfigs() updated

Add new test_securityConfig test
testing the full list of WPA3-CM parameters are
added to config map by the corresponding pwhm function
and the values are as expected (manual test with a sniffer)

Implements: PPM-3352

Signed-off-by: Iacob Juc <iacob.juc_ext@softathome.com>