pwhm: chanmgt: fix simultaneous switch of channel and bandwidth
Problem: In the 6GHz band, when switching from channel 213/320-2 to
149/320MHz-1, the bandwidth is changed to 160 instead of staying 320.

Cause: The channel and bandwidth data model handlers are called separately,
which leads pwhm to decide that 320MHz-1 is not available on channel 213, hence
changing to 160MHz, then only switching to channel 149.

Solution: Handle parameters that affect the target channel spec in a group
handler, i.e., in a single handler instead of separate handlers.

GitOrigin-RevId: a2cfab19982af5b6b91fc158d7bef6b03e4300c4

ExternalID: SSW-9943

References: PPW-308

Signed-off-by: Ahmed Amine HAOUES <ahmed-amine.haoues_ext@softathome.com>