pwhm: chanmgt: fix autoChannelselection still enabled when we change the channel from AP WebUI
Problem: AutoChannelEnable does not switch to False when the channel is selected manually

Cause: When ExternalAcsMgmt system is active, we shall read the pR->autoChannelSetByUser to determine if the AutoChannelEnable is set or not.

Solution:
Read the AutoChannelEnable datamodel value to avoid any configs order issue.

GitOrigin-RevId: 0bf1ef4bcd7140b22aa662c103a41b6ebe928f12

ExternalID: SSW-10037

References: PPW-308

Signed-off-by: Ahmed Amine HAOUES <ahmed-amine.haoues_ext@softathome.com>