pwhm: fsm : fix stations are not properly disconnected when hostapd is stopped
Problem: Station are not properly disconnected when hostapd is stopped.

Cause: No deauth are sent on the air when hostapd is stopped, deauth were sent out only when radio was explicitely disabled.

Solution: When hostapd daemon is stopped, deauth all stations on the radio.

GitOrigin-RevId: f6ea1ede80cc895f8f3b48434f54b83841206eb7

ExternalID: SSW-10354

Implements: PPM-3604

References: FEAT-125

Signed-off-by: Ahmed Amine HAOUES <ahmed-amine.haoues_ext@softathome.com>