pwhm: zwdfs: fix WiFi process is hanging when ZWDFS is launched
Problem: WiFi is hanging when starting ZWDFS when not supported.

Cause: Error status is not taken into account correctly which leads to a restart of ZWDFS when calling setChanspec.

Solution:

Return an error when ZWDFS is not available, in order to not go into the ZWDFS handler in a loop anymore in setChanspec function.
Also fix a log spam

GitOrigin-RevId: b3938f499eb34b24b9eed1a815bc2b14ca55cc5d

ExternalID: SSW-10605

Implements: PPM-3604

References: FEAT-125

Signed-off-by: Ahmed Amine HAOUES <ahmed-amine.haoues_ext@softathome.com>