pwhm: rad: fix radio status no up to date
Problem: Radio status in the datamodel was not being updated, and
corresponding events were not sent immediately after the socket with
hostapd was established and the manager became ready.

Cause: Radio status was not updated during handling of EndPoint and
AccessPoint.

Fix:
- Ensure radio status is updated when manager becomes ready
- Added logic to update radio status in both AccessPoint and EndPoint
handling paths.

GitOrigin-RevId: d12191c901863c677be71f5a6666d22b7e29e9c4

ExternalID: SSW-11336

Implements: PPM-3604

References: FEAT-125

Signed-off-by: Ahmed Amine HAOUES <ahmed-amine.haoues_ext@softathome.com>