pwhm: hostapd: fix Operating Class in 802.11k Beacon Request frames
Problem: The OperatingClass field is being sent with a value of zero.

Cause: The OperatingClass is currently in a local format. For 2.4GHz,
Operating Class 81 cannot be resolved because the frequency band is set
to NONE.

Fix: Use the global OperatingClass format for 802.11k requests, and
convert the local OperatingClass to its corresponding global value.

GitOrigin-RevId: 8cc163768dea485bed22de6b671de7fc43a83434

ExternalID: SSW-11632

Implements: PPM-3604

References: FEAT-125

Signed-off-by: Ahmed Amine HAOUES <ahmed-amine.haoues_ext@softathome.com>