pwhm: hostapd: prevent hidden SSIDs from being added to Reduced Neighbor Report (RNR)
Problem: Hidden SSIDs are incorrectly included in the RNR, exposing their
ShortSSID.
This ShortSSID can be matched in the ShortSSID List IE of a Probe Request,
causing the hidden BSS to respond with its clear SSID. This undermines the
privacy intent of hidden SSIDs.

Cause: BSSs with hidden SSIDs are not excluded from RNR generation,
leading to their presence in the RNR database maintained by hostapd.

Fix: Disable RRM Neighbor Report support for BSSs with hidden SSIDs.
This prevents them from being added to the RNR DB, ensuring ShortSSID is
not exposed and maintaining the hidden nature of the SSID.

GitOrigin-RevId: e05f3382ecc33861957c40a1475c85019a0fe85d

ExternalID: SSW-11431

Implements: PPM-3604

References: FEAT-125

Signed-off-by: Ahmed Amine HAOUES <ahmed-amine.haoues_ext@softathome.com>