pwhm: rad: fix pwhm random crash on restart.
Problem: pWhm crashes on restart randomly.

cause: the problem is due to pointing to non correct memory space because it was purged, and this memory space is used to call differed function.

Fix: when stopping pwhm, remove the callId of the differed call in order to cancel properly this differed call.

GitOrigin-RevId: a7cc5264c1125d13e8453ab25527ad27dd12a75e

ExternalID: SSW-11465

Implements: PPM-3604

References: FEAT-125

Signed-off-by: Ahmed Amine HAOUES <ahmed-amine.haoues_ext@softathome.com>