pwhm: assocDev:Issue: fix associatedDevice entry is not always created
Problem: AssociatedDevice entry is not always created

Cause: In case of parsing error, the whole AssociatedDevice entry is not created, while the station is successfully associated, which create a desynchro.

Solution: Even if parsing is failing, continue to process the station as a new AssociatedDevice.

GitOrigin-RevId: 467480d7c56e7b4f8051ee766635a0fa95347f9d

ExternalID: SSW-10020

Implements: PPM-3604

References: FEAT-125

Signed-off-by: Ahmed Amine HAOUES <ahmed-amine.haoues_ext@softathome.com>